refactored the blacklist impl
